% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfl_ss_unique}
\alias{dfl_ss_unique}
\title{The observation wells information conducted in Bulanghe Basin, China by Xi'an Geological Survey.}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{obs_well}{character, the observation well name of the 30 wells}
\item{X}{numeric, the x coordinate of the observation well in meter.}
\item{Y}{numeric, the y coordinate of the observation well in meter.}
\item{value}{numeric, the steady-state drawdown value in centi-meter.}
\item{pumping_test}{character, the pumping test name corresponding to the observation well.}
\item{element_number}{integer, the default element number in a 80\emph{80 element.}
\item{element_pump}{integer, the default element number for pumping well in a 80}80 element.}
}
}
\usage{
dfl_ss_unique
}
\description{
The observation wells information conducted in Bulanghe Basin, China by Xi'an Geological Survey.
}
\keyword{datasets}
