% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse.R
\name{Finverse}
\alias{Finverse}
\title{make a Monte Carlo inverse simulation.
HT simulations.
forward information. domain, bc, pump loc,
observation location and value.
geostatistical prior information.
hyper parameters --- nsim, itermax, etc.}
\usage{
Finverse(
  domain = c(40, 40, 0, 40, 0, 40),
  qHT = list(list(Qp = 10, xp = 20.5, yp = 20.5)),
  nsim = 50,
  itermax = 5,
  varmeanTmax = 5,
  rmsemin = 0,
  mul = 1,
  loc_obsHT,
  trueobshHT
)
}
\arguments{
\item{domain}{grid c(nx,ny,x1,x2,y1,y2)}

\item{qHT}{list of pumping test information.}

\item{nsim}{number of ensemble, default 50.}

\item{itermax}{maximum number of iterations.}

\item{varmeanTmax}{maximum variance of mean T.}

\item{rmsemin}{minimum rmse.}

\item{mul}{stablizer.}

\item{loc_obsHT}{list of observation location.}

\item{trueobshHT}{list of observation value.}
}
\value{
list of iteration data.
}
\description{
make a Monte Carlo inverse simulation.
HT simulations.
forward information. domain, bc, pump loc,
observation location and value.
geostatistical prior information.
hyper parameters --- nsim, itermax, etc.
}
\examples{
set.seed(100)
trueK <- random2d(nsim=1)
TT <- trueK[,-c(1,2)]
Qinf1=list(Qp=10,xp=20.5,yp=20.5)
trueh <- Fsteady2dsim(TT=TT,Qinf=Qinf1)
trueh <- trueh$solution
n <- 1600
nobs <- 25
locx = c(15,18,22,25,30)
locy = c(15,18,22,25,30)
loc= expand.grid(x=locx,y=locy)
loc_obs <- (loc$y-1)*40 + loc$x
trueobsh <- trueh[loc_obs]
loc_obsHT <- list(loc_obs)
trueobshHT <- list(trueobsh)
result <- Finverse(loc_obs=loc_obsHT,trueobsh=trueobshHT)
Qinf2=list(Qp=10,xp=10.5,yp=10.5)
trueh2 <- Fsteady2dsim(TT=TT,Qinf=Qinf2)
trueh2 <- trueh2$solution
trueobsh2 <- trueh2[loc_obs]
loc_obsHT <- list(loc_obs,loc_obs)
trueobshHT <- list(trueobsh,trueobsh2)
result <- Finverse(loc_obs=loc_obsHT,trueobsh=trueobshHT,qHT = list(Qinf1,Qinf2))
}
