% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getQseq3D}
\alias{getQseq3D}
\title{Map pumping well (x, y, z) coordinates to grid indices (3D)}
\usage{
getQseq3D(grid, Qinf = data.frame(Qp = 10, x = 20.5, y = 20.5, z = 5.5))
}
\arguments{
\item{grid}{grid generated by \code{GenGrid3D}.}

\item{Qinf}{a data frame with columns \code{Qp}, \code{x}, \code{y}, \code{z}.}
}
\value{
\code{Qinf} with additional columns \code{Nxp}, \code{Nyp}, \code{Nzp}.
}
\description{
Map pumping well (x, y, z) coordinates to grid indices (3D)
}
\examples{
grid3d <- GenGrid3D()
Qinf <- data.frame(Qp = 10, x = 20.5, y = 20.5, z = 5.5)
getQseq3D(grid = grid3d, Qinf = Qinf)
}
