% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oHT}
\alias{oHT}
\title{Observation data in list form}
\format{
A list of 12 data frames, each with columns:
\describe{
\item{data}{numeric, the observed drawdown \code{[L]}.}
\item{x}{numeric, the x coordinate of the observation well \code{[L]}.}
\item{y}{numeric, the y coordinate of the observation well \code{[L]}.}
}
}
\usage{
oHT
}
\description{
A list of 12 data frames, one per pumping test.  Each data frame contains
the observed drawdown and observation well location.  This is the format
expected by all forward and inverse functions.
}
\keyword{datasets}
