% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{Plotparameter2d}
\alias{Plotparameter2d}
\title{Plot the generated T distribution from random2d.}
\usage{
Plotparameter2d(
  TT,
  palette = "viridis",
  plotshow = TRUE,
  plotfile = NULL,
  plotwidth = 10,
  plotheight = 10,
  iflog = FALSE
)
}
\arguments{
\item{TT}{a data.frame, the result of Fsteady2dsim (x,y,transsimisivity).}

\item{palette}{a string, the palette to use. could be magma, inferno, plasma, viridis, cividis, rocket, marko, turbo, or simply A-G.}

\item{plotshow}{a logical value, if TRUE, the plot will be shown.}

\item{plotfile}{a string, the path to save the plot.}

\item{plotwidth}{a numeric value, the width of the plot.}

\item{plotheight}{a numeric value, the height of the plot.}

\item{iflog}{a logical value, if true,  use log(T), else using T.}
}
\value{
a plot showing contour.
}
\description{
Plot the generated T distribution from random2d.
}
\examples{
TT= random2d()
Plotparameter2d(TT,iflog=T)
TT= random2d(geo=list(me=0,var=1,geomod="Exp",anis=c(90,0.2),range=30,nugget=0))
Plotparameter2d(TT,iflog=T)
}
