% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{Plotsteady2d}
\alias{Plotsteady2d}
\title{Plot the generated head distribution from Fsteady2dsim.}
\usage{
Plotsteady2d(
  s,
  palette = "viridis",
  plotshow = TRUE,
  plotfile = NULL,
  plotwidth = 10,
  plotheight = 10,
  ifdrawdown = TRUE,
  Qinf = NULL,
  Oinf = NULL
)
}
\arguments{
\item{palette}{a string, the palette to use. could be magma, inferno, plasma, viridis, cividis, rocket, marko, turbo, or simply A-G.}

\item{plotshow}{a logical value, if TRUE, the plot will be shown.}

\item{plotfile}{a string, the path to save the plot.}

\item{plotwidth}{a numeric value, the width of the plot.}

\item{plotheight}{a numeric value, the height of the plot.}

\item{ifdrawdown}{a logical value, if true,  use drawdown, else using head change.}

\item{Qinf}{The pumping test information in a data.frame (Qp,xp,yp).}

\item{Oinf}{The observation information in a data.frame (data,x,y).}

\item{steady2d}{a data.frame, the result of Fsteady2dsim (x,y,headchange).}
}
\value{
a plot showing contour.
}
\description{
Plot the generated head distribution from Fsteady2dsim.
}
\examples{
Qinf = data.frame(Qp=c(10,10),x=c(10.5,30.5),y=c(20.5,20.5))
s <- Fsteady2dsim(Qinf=Qinf)
Plotsteady2d(s,Qinf=Qinf,Oinf=data.frame(data=NA,x=c(15.5,25.5),y=c(20.5,20.5)))
}
