% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumpingtest_information}
\alias{pumpingtest_information}
\title{Pumping test information from Bulanghe Basin, China}
\format{
A data frame with 12 rows and 7 variables:
\describe{
\item{test_name}{character, the pumping test name (e.g. \code{"A1-1"}).}
\item{pumping_well}{character, the pumping well name (A1 to A6).}
\item{Q_(L_per_s)}{numeric, the pumping rate in L/s \code{[L\eqn{^3}/T]}.}
\item{X}{numeric, the x coordinate of the pumping well \code{[L]}.}
\item{Y}{numeric, the y coordinate of the pumping well \code{[L]}.}
\item{start}{POSIXct, the start time of the pumping test.}
\item{end}{POSIXct, the end time of the pumping test.}
}
}
\usage{
pumpingtest_information
}
\description{
Details of 12 pumping tests conducted at the Bulanghe Basin by the Xi'an
Geological Survey, China.  Used as reference data for hydraulic tomography
inversion.
}
\keyword{datasets}
