% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qHT}
\alias{qHT}
\title{The pumping test information in package form.}
\format{
A list of 12 data frames, each data frame contains pumping rate and location.
\describe{
\item{A1-1}{data frame of pumping test 1, with columns Qp(m3/d), x, y.}
\item{A1-2}{data frame of pumping test 2, with columns Qp, x, y.}
...
}
}
\usage{
qHT
}
\description{
The pumping test information in package form.
}
\keyword{datasets}
