% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{samData}
\alias{samData}
\title{sample obs for synthetic data.}
\usage{
samData(
  Oinf = data.frame(data = NA, x = 10.5, y = 10.5),
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  h = rnorm(1600)
)
}
\arguments{
\item{Oinf}{a data.frame showing (data = null, x,y).}

\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}

\item{grid}{grid information generated by GenGrid function.}

\item{h}{the head field form forward simulation.}
}
\value{
output Oinf with data field filled.
}
\description{
sample obs for synthetic data.
}
\examples{
grid <- GenGrid()
Oinf=data.frame(data=NA,x=10.5,y=10.5)
h = rnorm(1600)
Oinf <- samData(Oinf=Oinf,grid=grid,h=h)
}
