% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{samData3D}
\alias{samData3D}
\title{Sample simulated head values at observation well locations (3D)}
\usage{
samData3D(Oinf = data.frame(data = NA, x = 10.5, y = 10.5, z = 5.5), grid, h)
}
\arguments{
\item{Oinf}{a data frame with columns \code{data}, \code{x}, \code{y}, \code{z}.}

\item{grid}{grid generated by \code{GenGrid3D}.}

\item{h}{the head vector from \code{Fsteady3dsim()$solution} (length \code{n}).}
}
\value{
\code{Oinf} with the \code{data} column filled.
}
\description{
Sample simulated head values at observation well locations (3D)
}
\examples{
grid3d <- GenGrid3D()
Oinf <- data.frame(data = NA, x = 10.5, y = 10.5, z = 5.5)
h <- rnorm(grid3d$n)
samData3D(Oinf = Oinf, grid = grid3d, h = h)
}
