% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient3D.R
\name{samData3DTr}
\alias{samData3DTr}
\title{Sample transient 3D head values at observation locations and times}
\usage{
samData3DTr(Oinf, grid, result_tr)
}
\arguments{
\item{Oinf}{data frame with columns \code{data}, \code{x}, \code{y},
\code{z}, \code{time}.}

\item{grid}{grid from \code{GenGrid3D()}.}

\item{result_tr}{list returned by \code{\link{Ftransient3dsim}}.}
}
\value{
\code{Oinf} with the \code{data} column filled.
}
\description{
Extracts simulated head values from the output of
\code{\link{Ftransient3dsim}} at the spatial locations and observation
times listed in \code{Oinf}.  The nearest available time step is used when
an observation time does not exactly match the ODE output.
}
\examples{
grid3d <- GenGrid3D(c(15, 15, 5, 0, 15, 0, 15, 0, 5))
times  <- c(0, 10, 50, 100)
res3d  <- Ftransient3dsim(grid = grid3d, times = times)
Oinf   <- data.frame(data = NA,
                     x    = c(5.5, 10.5),
                     y    = c(7.5,  7.5),
                     z    = c(2.5,  2.5),
                     time = c(50,  100))
Oinf   <- samData3DTr(Oinf = Oinf, grid = grid3d, result_tr = res3d)
}
