% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theis.R
\name{theis_drawdown}
\alias{theis_drawdown}
\title{Theis drawdown}
\usage{
theis_drawdown(Q, r, T, S, t)
}
\arguments{
\item{Q}{pumping rate \eqn{[L^3/T]} (positive = extraction).}

\item{r}{radial distance from pumping well \eqn{[L]}.}

\item{T}{aquifer transmissivity \eqn{[L^2/T]}.}

\item{S}{aquifer storativity \eqn{[-]}.}

\item{t}{numeric vector of elapsed times \eqn{[T]}.}
}
\value{
numeric vector of drawdown values \eqn{[L]}.
}
\description{
Compute transient drawdown in a confined aquifer using the Theis (1935)
solution.
}
\examples{
t <- 10^seq(1, 5, length.out = 100)
s <- theis_drawdown(Q = 1.3e-3, r = 200, T = 1.5e-3, S = 2e-5, t = t)
plot(t, s, type = "l", log = "x", xlab = "t", ylab = "s")
}
