% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady2D.R
\name{diffusion2D}
\alias{diffusion2D}
\title{2-D diffusion function solver}
\usage{
diffusion2D(t, h, par)
}
\arguments{
\item{t}{time}

\item{h}{initial condition (not important in steady-state solution)}

\item{par}{parameters}
}
\value{
list of state variables (e.g., head values)
}
\description{
2-D diffusion function solver
}
