% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse.R
\name{Finverse}
\alias{Finverse}
\title{make a Monte Carlo inverse simulation.
HT simulations.
forward information. domain, bc, pump loc,
observation location and value.
geostatistical prior information.
hyper parameters --- nsim, itermax, etc.}
\usage{
Finverse(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  qHT = list(data.frame(Qp = 10, x = 20.5, y = 20.5)),
  nsim = 50,
  itermax = 5,
  varmeanTmax = 5,
  rmsemin = 0,
  mul = 1,
  oHT = list(data.frame(data = -1, x = 11, y = 11)),
  lrw = 160000
)
}
\arguments{
\item{domain}{grid c(nx,ny,x1,x2,y1,y2)}

\item{grid}{grid generated by GenGrid function using domain information, should give if no domain.}

\item{qHT}{list of pumping test information.}

\item{nsim}{number of ensemble, default 50.}

\item{itermax}{maximum number of iterations.}

\item{varmeanTmax}{maximum variance of mean T.}

\item{rmsemin}{minimum rmse.}

\item{mul}{stablizer.}

\item{oHT}{list of observation information. list(data.frame(data,x,y))}

\item{lrw}{integer, the length of the real work array, default 160000 (grid size).}
}
\value{
list of iteration data.
}
\description{
make a Monte Carlo inverse simulation.
HT simulations.
forward information. domain, bc, pump loc,
observation location and value.
geostatistical prior information.
hyper parameters --- nsim, itermax, etc.
}
\examples{
set.seed(100)
trueK <- random2d(nsim=1)
TT <- trueK[,-c(1,2)]
domain=c(40,40,0,40,0,40)
grid = GenGrid(domain)
Qinf1=data.frame(Qp=10,xp=20.5,yp=20.5)
qHT <- list(test1 = Qinf1)
trueh <- Fsteady2dsim(TT=TT,Qinf=Qinf1,domain=domain)
locx = c(15,18,22,25,30)
locy = c(15,18,22,25,30)
loc= expand.grid(x=locx,y=locy)
Oinf1 <- data.frame(data=NA,x=loc$x,y=loc$y)
Oinf1 <- samData(grid = grid,Oinf = Oinf1,h=trueh$solution)
oHT <- list(test1 = Oinf1)
result <- Finverse(grid =grid, qHT = qHT, oHT = oHT)
inversePlot(niterm=5,iterdf = result,oHT = oHT,trueK=trueK)
Qinf2=data.frame(Qp=10,xp=10.5,yp=10.5)
trueh2 <- Fsteady2dsim(TT=TT,Qinf=Qinf2,domain=domain)
Oinf2 <- Oinf1
Oinf2 <- samData(grid = grid,Oinf = Oinf2,h=trueh2$solution)
oHT <- list(test1 = Oinf1,test2 = Oinf2)
qHT <- list(test1 = Qinf1,test2 = Qinf2)
result <- Finverse(grid =grid, qHT = qHT, oHT = oHT)
inversePlot(niterm=5,iterdf = result,oHT = oHT,trueK=trueK)
}
