% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady2D.R
\name{Fsteady2dsim}
\alias{Fsteady2dsim}
\title{Solve 2-D steady-state groundwater flow equation}
\usage{
Fsteady2dsim(
  domain = c(40, 40, 0, 40, 0, 40),
  TT = 0.1,
  Qinf = list(Qp = 10, xp = 20.5, yp = 20.5)
)
}
\arguments{
\item{domain}{a 6 * 1 vector giving (nx,ny,x1,x2,y1,y2)}

\item{TT}{a 1 * n vector giving transimisivity (L2/T) at each grid point.}

\item{Qinf}{a list or data.frame giving pumping rates (Qp) and locations (xp,yp).}
}
\value{
A data frame of two elements with grid showing generated grid and
steady-state solution of head distribution.
}
\description{
Solve 2-D steady-state groundwater flow equation. The boundary conditions at
both sides, the source term, and the parameters should be specified. A grid to
run the simulation should be given and parameters and source term rely on the
definition of grid.
}
\examples{
s <- Fsteady2dsim()
library('ggplot2')
ggplot(s) +
  aes(x = x, y = y, fill = solution) +
  geom_tile() +
  scale_fill_viridis_c(option = "viridis", direction = 1) +
  theme_minimal()
}
