% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady2D.R
\name{Fsteady2dsim}
\alias{Fsteady2dsim}
\title{Solve 2-D steady-state groundwater flow equation}
\usage{
Fsteady2dsim(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  TT = 0.1,
  Qinf = data.frame(Qp = 10, x = 20.5, y = 20.5),
  lrw = 160000
)
}
\arguments{
\item{domain}{a 6 * 1 vector giving (nx,ny,x1,x2,y1,y2)}

\item{grid}{grid generated by GenGrid function using domain information, should give if no domain.}

\item{TT}{a  n * 1 vector giving transimisivity (L2/T) at each grid point.}

\item{Qinf}{a ldata.frame giving pumping rates (Qp) and locations (x,y).}

\item{lrw}{integer, the length of the real work array, default 160000 (grid size).}
}
\value{
A data frame of two elements with grid showing generated grid and
steady-state solution of head distribution.
}
\description{
Solve 2-D steady-state groundwater flow equation. The boundary conditions at
both sides, the source term, and the parameters should be specified. A grid to
run the simulation should be given and parameters and source term rely on the
definition of grid.
}
\examples{
s <- Fsteady2dsim()
Plotsteady2d(s,plotfile="steady2d.pdf")
Qinf = data.frame(Qp=c(10,10),x=c(10.5,30.5),y=c(20.5,20.5))
s <- Fsteady2dsim(Qinf=Qinf)
Plotsteady2d(s)
set.seed(100)
TT <- random2d(nsim=1)$Tp
s <- Fsteady2dsim(TT=TT,Qinf=Qinf)
Plotsteady2d(s)
}
