% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenGrid}
\alias{GenGrid}
\title{generate forward running grid and output its grid and observation/source grid.
input domain, pump and obs location}
\usage{
GenGrid(
  domain = c(40, 40, 0, 40, 0, 40),
  pump = list(xp = 20.5, yp = 20.5),
  obs = list(xo = 30, yo = 30)
)
}
\arguments{
\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}

\item{pump}{c(xp,yp) pump location.}

\item{obs}{c(xo,yo) observation location.}
}
\value{
output grid, pump and obs element number
}
\description{
generate forward running grid and output its grid and observation/source grid.
input domain, pump and obs location
}
\examples{
grid <- GenGrid()
}
