% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenGrid}
\alias{GenGrid}
\title{generate forward running grid and output its grid and grid.
input domain}
\usage{
GenGrid(domain = c(40, 40, 0, 40, 0, 40))
}
\arguments{
\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}
}
\value{
output n, xmid, ymid, and grid
}
\description{
generate forward running grid and output its grid and grid.
input domain
}
\examples{
grid <- GenGrid()
}
