% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getOelem}
\alias{getOelem}
\title{Get the elemental number of the observation well.}
\usage{
getOelem(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  Oinf = data.frame(data = NA, x = 30.5, y = 30.5)
)
}
\arguments{
\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}

\item{grid}{grid information generated by GenGrid function.}

\item{Oinf}{a data.frame showing (data = null, x,y)}
}
\value{
output elemental number of the observation wells.
}
\description{
Get the elemental number of the observation well.
}
\examples{
grid <- GenGrid()
Oinf=data.frame(data=NA,x=30.5,y=30.5)
Oinf <- getOelem(grid = grid, Oinf=Oinf)

}
