% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getQseq}
\alias{getQseq}
\title{Get the x- and y- directional sequence of the pumping well.}
\usage{
getQseq(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  Qinf = data.frame(Qp = 10, x = 20.5, y = 20.5)
)
}
\arguments{
\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}

\item{grid}{grid information generated by GenGrid function.}

\item{Qinf}{a data.frame showing (Qp,x,y)}
}
\value{
output x- and y- directional sequence Nxp and Nyp
}
\description{
Get the x- and y- directional sequence of the pumping well.
}
\examples{
grid <- GenGrid()
Qseq <- getQseq(grid = grid)

}
