% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{inversePlot}
\alias{inversePlot}
\title{now we can choose the best iteration to plot.
estimated K plot.}
\usage{
inversePlot(niterm = 1, iterdf, trueh, trueK = NULL)
}
\arguments{
\item{iterdf}{the data.frame of iterations.}

\item{trueK}{the true K field (used in synthetic case).}

\item{niter}{the number of iterations.}
}
\description{
now we can choose the best iteration to plot.
estimated K plot.
}
\examples{
set.seed(100)
trueK <- random2d(nsim=1)
TT <- trueK[,-c(1,2)]
trueh <- Fsteady2dsim(TT=TT)
trueh <- trueh$solution
n <- 1600
nobs <- 25
locx = c(15,18,22,25,30)
locy = c(15,18,22,25,30)
loc= expand.grid(x=locx,y=locy)
loc_obs <- (loc$y-1)*40 + loc$x
trueobsh <- trueh[loc_obs]
loc_obs = list(loc_obs)
trueobsh = list(trueobsh)
result <- Finverse(loc_obs=loc_obs,trueobsh=trueobsh)
inversePlot(niterm=5,iterdf = result,trueh = trueobsh,trueK=trueK)
}
