% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{inversePlot}
\alias{inversePlot}
\title{now we can choose the best iteration to plot.}
\usage{
inversePlot(niterm = 1, grid, iterdf, oHT, trueK = NULL)
}
\arguments{
\item{niterm}{the number of iterations.}

\item{grid}{grid generated by GenGrid function using domain information.}

\item{iterdf}{the data.frame of iterations.}

\item{oHT}{list of observation information. list(data.frame(data,x,y))}

\item{trueK}{the true K field (used in synthetic case).}
}
\description{
now we can choose the best iteration to plot.
}
\examples{
set.seed(100)
trueK <- random2d(nsim=1)
TT <- trueK[,-c(1,2)]
domain=c(40,40,0,40,0,40)
grid = GenGrid(domain)
Qinf1=data.frame(Qp=10,xp=20.5,yp=20.5)
qHT <- list(test1 = Qinf1)
trueh <- Fsteady2dsim(TT=TT,Qinf=Qinf1,grid=grid)
trueh <- trueh$solution
locx = c(15,18,22,25,30)
locy = c(15,18,22,25,30)
loc= expand.grid(x=locx,y=locy)
Oinf <- data.frame(data=NA,x=loc$x,y=loc$y)
Oinf <- getOelem(grid = grid)
Oinf$data <- trueh[Oinf$nelem]
oHT <- list(test1 = Oinf)
result <- Finverse(grid =grid, qHT = qHT, oHT = oHT)
inversePlot(niterm=5,grid=grid,iterdf = result,oHT = oHT,trueK=trueK)
}
