% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{inversePlot}
\alias{inversePlot}
\title{now we can choose the best iteration to plot.}
\usage{
inversePlot(
  niterm = 1,
  grid,
  iterdf,
  oHT,
  trueK = NULL,
  p1title = "Est. lnT value",
  p1z = "lnT",
  p1xlab = "X/m",
  p1ylab = "Y/m",
  p2title = "Est. var. of lnT",
  p2z = "var. of lnT",
  p2xlab = "X/m",
  p2ylab = "Y/m",
  p3title = "Drawdown scatter",
  p3xlab = "Obs. Drawdown/m",
  p3ylab = "Sim. Drawdown/m",
  p4title = "lnT scatter",
  p4xlab = "True lnT /(m2/d)",
  p4ylab = "Est. lnT /(m2/d)"
)
}
\arguments{
\item{niterm}{the number of iterations.}

\item{grid}{grid generated by GenGrid function using domain information.}

\item{iterdf}{the data.frame of iterations.}

\item{oHT}{list of observation information. list(data.frame(data,x,y))}

\item{trueK}{the true K field (used in synthetic case).}

\item{p1title}{title of plot 1 (value of lnT).}

\item{p1z}{z label of plot 1.}

\item{p1xlab}{x label of plot 1.}

\item{p1ylab}{y label of plot 1.}

\item{p2title}{title of plot 2 (value of lnT variance).}

\item{p2z}{z label of plot 2.}

\item{p2xlab}{x label of plot 2.}

\item{p2ylab}{y label of plot 2.}

\item{p3title}{title of plot 3(head scatter).}

\item{p3xlab}{x label of plot 3.}

\item{p3ylab}{y label of plot 3.}

\item{p4title}{title of plot 4(lnT scatter).}

\item{p4xlab}{x label of plot 4.}

\item{p4ylab}{y label of plot 4.}
}
\description{
now we can choose the best iteration to plot.
}
\examples{
set.seed(100)
trueK <- random2d(nsim=1)
TT <- trueK[,-c(1,2)]
domain=c(40,40,0,40,0,40)
grid = GenGrid(domain)
Qinf1=data.frame(Qp=10,xp=20.5,yp=20.5)
qHT <- list(test1 = Qinf1)
trueh <- Fsteady2dsim(TT=TT,Qinf=Qinf1,grid=grid)
trueh <- trueh$solution
locx = c(15,18,22,25,30)
locy = c(15,18,22,25,30)
loc= expand.grid(x=locx,y=locy)
Oinf <- data.frame(data=NA,x=loc$x,y=loc$y)
Oinf <- getOelem(grid = grid)
Oinf$data <- trueh[Oinf$nelem]
oHT <- list(test1 = Oinf)
result <- Finverse(grid =grid, qHT = qHT, oHT = oHT)
inversePlot(niterm=5,grid=grid,iterdf = result,oHT = oHT,trueK=trueK)
}
