% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oHT}
\alias{oHT}
\title{The pumping test information in package form.}
\format{
A list of 12 data frames, each data frame contains observation data and location.
\describe{
\item{A1-1}{data frame of pumping test 1, with columns data(drawdown in m), x, y.}
\item{A1-2}{data frame of pumping test 2, with columns data, x, y.}
\item{...}{...}
}
}
\usage{
oHT
}
\description{
The pumping test information in package form.
}
\keyword{datasets}
