% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroPlot.R
\name{Plotsteady2d}
\alias{Plotsteady2d}
\title{Plot the generated head distribution from Fsteady2dsim.}
\usage{
Plotsteady2d(
  s,
  palette = "viridis",
  plotshow = TRUE,
  plotfile = NULL,
  plotwidth = 10,
  plotheight = 10,
  ifdrawdown = TRUE
)
}
\arguments{
\item{palette}{a string, the palette to use. could be magma, inferno, plasma, viridis, cividis, rocket, marko, turbo, or simply A-G.}

\item{plotshow}{a logical value, if TRUE, the plot will be shown.}

\item{plotfile}{a string, the path to save the plot.}

\item{plotwidth}{a numeric value, the width of the plot.}

\item{plotheight}{a numeric value, the height of the plot.}

\item{ifdrawdown}{a logical value, if true,  use drawdown, else using head change.}

\item{steady2d}{a data.frame, the result of Fsteady2dsim (x,y,headchange).}
}
\value{
a plot showing contour.
}
\description{
Plot the generated head distribution from Fsteady2dsim.
}
