% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qHT}
\alias{qHT}
\title{Pumping test information in list form}
\format{
A list of 12 data frames, each with columns:
\describe{
\item{Qp}{numeric, the pumping rate in m\eqn{^3}/d \code{[L\eqn{^3}/T]}.}
\item{x}{numeric, the x coordinate of the pumping well \code{[L]}.}
\item{y}{numeric, the y coordinate of the pumping well \code{[L]}.}
}
}
\usage{
qHT
}
\description{
A list of 12 data frames, one per pumping test.  Each data frame contains
the pumping rate and well location.  This is the format expected by all
forward and inverse functions.
}
\keyword{datasets}
