% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{random2d}
\alias{random2d}
\title{Generate 2-D random field of hydraulic parameters}
\usage{
random2d(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  geo = list(me = 0, var = 1, geomod = "Exp", anis = c(90, 1), range = 30, nugget = 0),
  nsim = 1
)
}
\arguments{
\item{domain}{c(Nx,Ny,x1,x2,y1,y2) domain of the grid.}

\item{geo}{geostatistical features: list(mean,variance,covarirance_function,anisotropy,range,nugget)}

\item{nsim}{number of simulations.}
}
\value{
A data frame of three elements with grid showing generated grid (in x and y) and
parameter values.
}
\description{
Based on package gstat, it generate correlated random field using Gaussian simulation.
}
\examples{
TT= random2d()
Plotparameter2d(TT,iflog=T)
 TT= random2d(geo=list(me=0,var=1,geomod="Exp",anis=c(90,0.2),range=30,nugget=0))
Plotparameter2d(TT,iflog=T)
}
