% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{samData}
\alias{samData}
\title{sample obs for synthetic data.}
\usage{
samData(loc_obs = c(1, 2), h = rnorm(10))
}
\arguments{
\item{loc_obs}{c(elem1,elem2) elem number of the observation}
}
\description{
sample obs for synthetic data.
}
