% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient2D.R
\name{samDataTr}
\alias{samDataTr}
\title{Sample transient 2D head values at observation locations and times}
\usage{
samDataTr(Oinf, grid, result_tr)
}
\arguments{
\item{Oinf}{data frame with columns \code{data}, \code{x}, \code{y},
\code{time}.}

\item{grid}{grid from \code{GenGrid()}.}

\item{result_tr}{list returned by \code{\link{Ftransient2dsim}}.}
}
\value{
\code{Oinf} with the \code{data} column filled.
}
\description{
Extracts simulated head values from the output of
\code{\link{Ftransient2dsim}} at the spatial locations and observation
times listed in \code{Oinf}.
}
\details{
If an observation time does not exactly match a row in \code{result_tr},
the nearest available time step is used.  To get exact values, include all
observation times in the \code{times} argument of
\code{\link{Ftransient2dsim}}.
}
\examples{
times  <- seq(0, 100, by = 10)
res    <- Ftransient2dsim(times = times)
Oinf   <- data.frame(data = NA,
                     x    = c(10.5, 25.5, 30.5),
                     y    = c(20.5, 15.5, 25.5),
                     time = c(10,   50,   100))
Oinf   <- samDataTr(Oinf = Oinf, grid = res$grid, result_tr = res)
}
