% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inverse.R
\name{samHTmc}
\alias{samHTmc}
\title{This function takes qHT and oHT as input and return sampled data in monte Carlo simulation.}
\usage{
samHTmc(
  domain = c(40, 40, 0, 40, 0, 40),
  grid = NULL,
  TT = 0.1,
  qHT = list(data.frame(Qp = 10, x = 20.5, y = 20.5)),
  oHT = list(data.frame(data = -1, x = 11, y = 11)),
  lrw = 160000
)
}
\arguments{
\item{domain}{a 6 * 1 vector giving (nx,ny,x1,x2,y1,y2)}

\item{grid}{grid generated by GenGrid function using domain information, should give if no domain.}

\item{TT}{a  n * 1 vector giving transimisivity (L2/T) at each grid point.}

\item{qHT}{pumping test information.}

\item{oHT}{observation information.}

\item{lrw}{integer, the length of the real work array, default 160000 (grid size).}
}
\value{
Updated oHT or just one vector given the obs data.
}
\description{
This function takes qHT and oHT as input and return sampled data in monte Carlo simulation.
}
\examples{
domain=c(80,80,0,80,0,80)
grid <- GenGrid(domain=domain)
TT= random2d(nsim=10,grid=grid)
TT = TT[,-c(1,2)]
data("oHT")
data("qHT")
da = samHTmc(grid =grid, TT=TT, qHT= qHT, oHT=oHT,lrw=320000)
}
